requirements.txt
prettyprint
getpass4
pyyaml
tabulate
icecream